outsourced LabelEditingPreview
