+++ app/lib/resourcemodeler/features/label-editing/index.js
@@ -3,7 +3,7 @@ import ResizeModule from 'diagram-js/lib/features/resize';
 import DirectEditingModule from 'diagram-js-direct-editing';
 
 import LabelEditingProvider from './LabelEditingProvider';
-import LabelEditingPreview from './LabelEditingPreview';
+import LabelEditingPreview from '../../../frequentlyUsedFiles/features/label-editing/LabelEditingPreview.js';
 
 
 export default {
